Added the files to correct places
